Moving File to lib/mumuki/laboratory
