+++ lib/mumuki/laboratory.rb
@@ -28,3 +28,4 @@ require_relative './laboratory/exceptions'
 require_relative './laboratory/status'
 require_relative './laboratory/evaluation'
 require_relative './laboratory/controllers'
+require_relative './laboratory/file'
