+++ lib/mumuki/laboratory/file.rb
@@ -1,0 +1,18 @@
+module Mumuki::Laboratory
+  class File
+    attr_reader :name, :content
+
+    def initialize(name, content)
+      @name = name
+      @content = content
+    end
+
+    def highlight_mode
+      Language.find_by(extension: extension)&.highlight_mode || extension
+    end
+
+    def extension
+      name.get_file_extension
+    end
+  end
+end
